Rename table module to models.
