+++ orangecontrib/datafusion/widgets/owmeanfuser.py
@@ -5,7 +5,7 @@ from PyQt4 import QtGui, QtCore
 from Orange.widgets import widget, gui, settings
 
 from skfusion import fusion
-from orangecontrib.datafusion.table import Relation
+from orangecontrib.datafusion.models import Relation
 from orangecontrib.datafusion.widgets.owfusiongraph import \
     SimpleTableWidget, rel_shape, rel_cols, RelationCompleter
