+++ orangecontrib/datafusion/widgets/owmoviegenres.py
@@ -5,7 +5,7 @@ from PyQt4 import QtGui
 from PyQt4.QtGui import QGridLayout
 from Orange.widgets.widget import OWWidget
 from Orange.widgets import widget, gui
-from orangecontrib.datafusion.table import Relation
+from orangecontrib.datafusion.models import Relation
 
 from skfusion import fusion
