+++ orangecontrib/datafusion/widgets/owmovieratings.py
@@ -2,7 +2,7 @@ import sys
 from PyQt4 import QtGui
 from Orange.widgets.widget import OWWidget
 from Orange.widgets import widget, gui, settings
-from orangecontrib.datafusion.table import Relation
+from orangecontrib.datafusion.models import Relation
 from orangecontrib.datafusion import movielens
 
 from skfusion import fusion
